'on the 8 '

+++ 0x07-pointers_arrays_strings/8-print_diagsums.c
@@ -10,15 +10,16 @@
 void print_diagsums(int *a, int size)
 {
 	int i;
-	int sum1, sum2;
+	long int sum1, sum2;
 
 	sum1 = 0;
 	sum2 = 0;
+	
 	for (i = 0 ; i < size ; i++)
 	{
 		sum1 += *(a + i * size + i);
 		sum2 += *(a + i * size + (size - i - 1));
 	}
-	printf("%d, %d \n", sum1, sum2);
+	printf("%ld, %ld \n", sum1, sum2);
 
 }
